[PT Run] Refactoring: moved common files to Plugins\Common and added references to plugins that use them
