+++ src/modules/launcher/Plugins/Microsoft.Plugin.Program/Programs/UWPApplication.cs
@@ -20,7 +20,6 @@ using System.Windows.Media.Imaging;
 using System.Xml;
 using ManagedCommon;
 using Microsoft.Plugin.Program.Logger;
-using Microsoft.Plugin.Program.Win32;
 using Wox.Infrastructure;
 using Wox.Infrastructure.Image;
 using Wox.Plugin;
@@ -331,7 +330,7 @@ namespace Microsoft.Plugin.Program.Programs
                 string source = $"@{{{packageFullName}? {parsed}}}";
                 var capacity = (uint)outBuffer.Capacity;
                 var hResult = NativeMethods.SHLoadIndirectString(source, outBuffer, capacity, IntPtr.Zero);
-                if (hResult == Hresult.Ok)
+                if (hResult == NativeMethods.HRESULT.S_OK)
                 {
                     var loaded = outBuffer.ToString();
                     if (!string.IsNullOrEmpty(loaded))
