--- src/modules/launcher/Plugins/Microsoft.Plugin.Uri/Interfaces/IRegistryWrapper.cs
@@ -1,11 +1,0 @@
-// Copyright (c) Microsoft Corporation
-// The Microsoft Corporation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-namespace Microsoft.Plugin.Uri.Interfaces
-{
-    public interface IRegistryWrapper
-    {
-        string GetRegistryValue(string registryLocation, string valueName);
-    }
-}
