+++ src/modules/launcher/Plugins/Microsoft.Plugin.Uri/Microsoft.Plugin.Uri.csproj
@@ -109,4 +109,13 @@
     </None>
   </ItemGroup>
 
+  <ItemGroup>
+    <Compile Include="..\Common\NativeMethods.cs">
+      <Link>NativeMethods.cs</Link>
+    </Compile>
+    <Compile Include="..\Common\BrowserInfo.cs">
+      <Link>BrowserInfo.cs</Link>
+    </Compile>
+  </ItemGroup>
+  
 </Project>
