--- src/modules/launcher/Plugins/Microsoft.Plugin.Uri/NativeMethods.cs
@@ -1,21 +1,0 @@
-// Copyright (c) Microsoft Corporation
-// The Microsoft Corporation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System;
-using System.Runtime.InteropServices;
-using System.Text;
-
-namespace Microsoft.Plugin.Uri
-{
-    internal static class NativeMethods
-    {
-        internal enum Hresult : uint
-        {
-            Ok = 0x0000,
-        }
-
-        [DllImport("shlwapi.dll", CharSet = CharSet.Unicode)]
-        internal static extern Hresult SHLoadIndirectString(string pszSource, StringBuilder pszOutBuf, uint cchOutBuf, IntPtr ppvReserved);
-    }
-}
