+++ src/modules/launcher/Plugins/Microsoft.Plugin.WindowWalker/Microsoft.Plugin.WindowWalker.csproj
@@ -113,4 +113,10 @@
     </None>
   </ItemGroup>
 
+  <ItemGroup>
+    <Compile Include="..\Common\NativeMethods.cs">
+      <Link>NativeMethods.cs</Link>
+    </Compile>
+  </ItemGroup>
+
 </Project>
