+++ src/modules/launcher/Plugins/Microsoft.PowerToys.Run.Plugin.System/Microsoft.PowerToys.Run.Plugin.System.csproj
@@ -139,5 +139,11 @@
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
   </ItemGroup>
+  
+  <ItemGroup>
+    <Compile Include="..\Common\NativeMethods.cs">
+      <Link>NativeMethods.cs</Link>
+    </Compile>
+  </ItemGroup>
 
 </Project>
